Fixing "more" bug and a bunch of UI freezes
the "more" bug was that if you had a split configured with "more", which is supposed to work sort of like "and," it was checking the address provided in the parent split instead of the "more" split.

the "connection ready" checking was doing a websocket "ping" every time "update splits" was called, which is every 33ms. this caused MASSIVE UI freezing, and also tons and tons of periods where the splitter couldn't get around to actually checking if it was time to split because it was busy doing connection checks. This removes most of those and does reconnect attempts when data or connections are dropped.

also various other stuff - some refactoring for style, some dropping of random array allocations, removing some unused members, improved "logging", fixing a crash when the connection to USB2SNES gets closed.