Fix handling of failed connection
Previously, when USB2SnesW.Connect was called, if the connection failed,
the websocket would be closed, but the TaskCompletionSource created by
Connect would never be set, so the caller would never get the result.
Additionally, the OnOpen and OnError handlers that were added to listen
for the connection were never removed.

This change adds a listener class that properly handles OnClose in
addition to OnOpen and OnError, and cleans up all of the handlers when
it's done.