Slightly better handling of size
Removed the code that tried to force the timer form to stay the same
size when the autosplitter component's size changed, because that didn't
work if you weren't also using a Splits component that could absorb the
size changes. Now, we delay showing the status message for a bit to give
the timer form a chance to settle into the right size before we start
changing the component's size. This is still ugly, but at least it works
slightly better than before.

Also removed some very spammy debug logs.